requirements.txt
ccxt==4.2.77
pandas==2.0.3
numpy==1.26.4
python-dotenv==1.0.1
backtrader==1.9.78.123
pandas_ta==0.3.14b0
# TA-Lib for non-Windows systems
TA-Lib==0.4.28; sys_platform != 'win32'

# Windows 사용자를 위한 TA-Lib 설치 참고
# pip install TA-Lib 시 아래 공식 사이트에서 whl 파일 다운로드 필요
# https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib